+++ CHANGELOG.md
@@ -7,13 +7,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [v5.1.0] - 2020-01-06
+
+### Added
+
+*   `StatusBar` with message, progress bar and optional left and right items.
+*   `AutoCompleteTextField` with customizable `SuggestionProvider`.   
+
 ## [v5.0.0] - 2019-12-19
 
 ### Added
 
 *   `Hint` a specialized tooltip showing a hint for an node, like a text field.
-*   Re-added `StatusBar` with message, progress bar and optional left and right
-    items.
 
 ### Removed